### 2024-12-16 #224 @C4dynamics
c4dynamics
Submitting Author: Ziv (@c4dynamics)
Package Name: c4dynamics
One-Line Description of Package: Python framework for algorithms of dynamic systems

Repository Link (if existing): https://github.com/C4dynamics/C4dynamics

---

c4dynamics is designed to simplify the development of algorithms for dynamic systems, using state space representations. It offers engineers and researchers a systematic approach to model, simulate, and control systems in fields like robotics, aerospace, and navigation.
The framework introduces state objects, which are foundational data structures that encapsulate state vectors and provide the tools for managing data, simulating system behavior, and analyzing results.
With integrated modules for sensors, detectors, and filters, c4dynamics accelerates algorithm development while maintaining flexibility and scalability.

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

- [ ] Astropy: [My package adheres to Astropy community standards](https://www.pyopensci.org/software-peer-review/partners/astropy.html)

None of these. c4dynamics belongs to the category of math-operations and data-operations for physical models.

---

- Explain how and why the package falls under these categories (briefly, 1-2 sentences). For community partnerships, check also their specific guidelines as documented in the links above. Please note any areas you are unsure of:
c4dynamics belongs to the category of math-operations and data-operations for physical models. Please refer to this page from the documentary to a brief overview of the package core: https://c4dynamics.github.io/C4dynamics/api/States.html

Researchers, engineers, and students from the fields of aerospace, robotics, navigation guidance and control.

The core concept of state objects - probably not.
Modules from the library of the package, like Kalman filters, seeker models, solvers etc. - there are. However, here they are built on top of a state object, which according to the philosophy of the package, should facilitate and improve the algorithm development.

I'd be happy if the kind reviewer contact with me for any question or uncertainty. I'm open to any kind of communication.

thank you :-)

[Conduct]: https://www.pyopensci.org/handbook/CODE_OF_CONDUCT.html
