+++ paper/paper.md
@@ -1,0 +1,206 @@
+---
+title: 'C4DYNAMICS: Python Framework for Dynamic Systems'
+tags:
+  - Python
+  - dynamics
+  - state space
+  - state objects 
+  - filters
+  - sensors
+  - detectors
+authors:
+  - name: Ziv Meri
+    orcid: 0000-0002-7932-0562
+date: December 17 2024
+---
+
+# Abstract
+
+<div style="text-align: justify">  
+
+Dynamic systems play a critical role across various fields such as robotics, aerospace, and control theory. While Python offers robust mathematical tools, it lacks a dedicated framework tailored for dynamic systems. C4dynamics addresses this gap by introducing a Python-based platform designed for state-space modeling and analysis. With the `state object` at its core, c4dynamics simplifies the development of algorithms for sensors, filters, and detectors through its modular structure. The framework's state objects, along with an integrated scientific library, empower researchers, engineers, and students to effectively develop algorithms for dynamic systems.
+
+
+# Introduction
+Dynamic systems are central to a wide range of scientific and engineering disciplines, from `control theory` to `robotics` and `aerospace`. These systems involve changing states over time, typically described by differential or difference equations. 
+Unlike static systems, dynamic systems evolve with time, making them more complex to model and analyze. The `state-space model` formalizes this behavior by representing the system's state as a vector of variables. This vector evolves according to predefined rules and provides a structured approach to modeling, making it a natural fit for the analysis of dynamic systems.
+The general form of the state-space model is:
+
+$$
+x[k+1]=A⋅x[k]+B⋅u[k]  
+$$
+$$
+y[k]=C⋅x[k]  
+$$
+Where:  
+$x[k]$ is the state vector  
+$u[k]$ is the input  
+$y[k]$ is the output  
+$A,B,C$ are matrices defining the system dynamics  
+$k$ is the discrete-time parameter  
+ 
+State-space models are widely used to represent both linear and nonlinear systems. 
+
+`MATLAB` has long been a popular choice for dynamic systems modeling due to its strong support for matrix operations and built-in toolboxes. However, `Python` offers distinct advantages, such as open-source accessibility, a large community, and integration with modern software engineering practices. 
+Python’s ecosystem, with libraries like NumPy and SciPy, provides robust mathematical tools, but it lacks a dedicated framework specifically for dynamic systems. 
+
+`C4DYNAMICS` bridges this gap, offering a flexible environment for developing and implementing algorithms that handle dynamic systems using the state-space representation in Python. 
+Designed to simplify the development of algorithms, c4dynamics offers engineers and researchers a systematic approach to model, simulate, and control systems in fields like robotics, aerospace, and navigation.
+With integrated modules for sensors, detectors, and filters, c4dynamics accelerates algorithm development while maintaining flexibility and scalability.
+
+
+# C4dynamics Architecture
+C4dynamics is structured around two primary building blocks:
+
+## State Data Structures
+These classes encapsulate a state vector, allowing data management and mathematical operations on the system state. 
+
+## Scientific Library
+Modules in this library operate on state objects, enabling the development of algorithms for sensors, filters, and detectors.
+
+![alt text](dependency.png)
+figure 1: C4dynamics dependency diagram 
+
[…]
