### 2024-12-19 cc81425
remove unnecessary varying_q.png, ballistics.rst
