+++ docs/source/programs/ballisticestimation.rst
@@ -1,11 +1,0 @@
-********************************
-Ballistic Coefficient Estimation
-********************************
-
-
-An estimation of the ballistic coefficient of a vertically falling target tracked by an altitude radar. 
-
-
-
-
-
