### 2024-12-20 4a01709
update bug in ekf example in filters intro
