+++ c4dynamics/__init__.py
@@ -79,7 +79,7 @@ from . import detectors
 #
 # version
 ##
-__version__ = '2.0.1'
+__version__ = '2.0.3'
 
 
 #
