+++ docs/html/_sources/api/generated/dp/c4dynamics.states.lib.datapoint.datapoint.inteqm.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.lib.datapoint.datapoint.inteqm
+================================================
+
+.. currentmodule:: c4dynamics.states.lib.datapoint
+
+.. automethod:: datapoint.inteqm
