+++ docs/html/_sources/api/States.rst
@@ -1,0 +1,32 @@
+State objects 
+=============
+
+
+.. automodule:: c4dynamics.states
+
+
+.. _objects-header:
+
+Objects 
+-------
+
+.. toctree::
+  :hidden:
+  :maxdepth: 1
+
+  The state class <states.state> 
+  States library <states.lib>
+
+
+
+.. list-table:: 
+  :header-rows: 0
+
+  * - `State <states.state>`_
+    - The state class.
+  * - `States library <states.lib>`_
+    - A collection of predefined states.
+
+
+
+
