+++ docs/html/_sources/gettingstarted.rst
@@ -1,0 +1,40 @@
+Getting Started
+===============
+
+
+1. **Install the Framework**
+
+   You can install c4dynamics via pip:
+   
+   ``pip install c4dynamics``
+
+
+2. **Explore the State Data Structure**
+
+   Learn how to define and manipulate states, 
+   the core units of dynamic systems within the framework.
+
+
+3. **Utilize Functional Libraries**
+
+   Incorporate sensors, detectors, and filters to build real-world systems.
+
+
+4. **Follow the Workflow** 
+
+   Deepen your understanding by reviewing the workflow outline:
+
+   .. figure:: /_architecture/workflow.png 
+
+
+5. **User Guide** 
+
+   Refer to the user guide to explore 
+   specific use cases and examples.
+
+
+6. **API Reference**
+
+   Review the comprehensive explanations and examples in the API documentation.
+
+
