+++ docs/html/_sources/api/generated/rb/c4dynamics.states.lib.rigidbody.rigidbody.RB.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.lib.rigidbody.rigidbody.RB
+============================================
+
+.. currentmodule:: c4dynamics.states.lib.rigidbody
+
+.. autoproperty:: rigidbody.RB
