+++ docs/html/_sources/api/generated/fdp/c4dynamics.states.lib.pixelpoint.pixelpoint.box.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.lib.pixelpoint.pixelpoint.box
+===============================================
+
+.. currentmodule:: c4dynamics.states.lib.pixelpoint
+
+.. autoproperty:: pixelpoint.box
