+++ docs/html/_sources/api/generated/dp/c4dynamics.states.lib.datapoint.datapoint.plot.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.lib.datapoint.datapoint.plot
+==============================================
+
+.. currentmodule:: c4dynamics.states.lib.datapoint
+
+.. automethod:: datapoint.plot
