+++ docs/html/_sources/api/generated/rotmat/c4dynamics.rotmat.rotmat.rotz.rst
@@ -1,0 +1,6 @@
+c4dynamics.rotmat.rotmat.rotz
+=============================
+
+.. currentmodule:: c4dynamics.rotmat.rotmat
+
+.. autofunction:: rotz
