+++ docs/html/_sources/api/generated/yolov3/c4dynamics.detectors.yolo3_opencv.yolov3.confidence_th.rst
@@ -1,0 +1,6 @@
+c4dynamics.detectors.yolo3\_opencv.yolov3.confidence\_th
+========================================================
+
+.. currentmodule:: c4dynamics.detectors.yolo3_opencv
+
+.. autoproperty:: yolov3.confidence_th
