+++ docs/html/_sources/programs/index.rst
@@ -1,0 +1,24 @@
+Programs
+========
+
+
+.. toctree::
+   :caption: Getting Started
+   :hidden:
+   :maxdepth: 1
+
+   dof6sim.ipynb
+
+
+
+Selected examples that demonstrate the framework's versatility, showcasing its use in diverse fields. 
+
+
+.. list-table:: 
+  :header-rows: 0
+
+  * - `6DOF <dof6sim.ipynb>`_
+    - Six degrees of freedom simulation of a missile pursuing a target.
+
+
+
