+++ docs/html/_sources/api/generated/state/c4dynamics.states.state.state.normalize.rst
@@ -1,6 +1,0 @@
-c4dynamics.states.state.state.normalize
-=======================================
-
-.. currentmodule:: c4dynamics.states.state
-
-.. autoproperty:: state.normalize
