+++ docs/html/_sources/api/generated/rb/c4dynamics.states.lib.rigidbody.rigidbody.mass.rst
@@ -1,6 +1,0 @@
-c4dynamics.states.lib.rigidbody.rigidbody.mass
-==============================================
-
-.. currentmodule:: c4dynamics.states.lib.rigidbody
-
-.. autoproperty:: rigidbody.mass
