+++ docs/html/_sources/api/generated/seeker/c4dynamics.sensors.seeker.seeker.bias.rst
@@ -1,6 +1,0 @@
-c4dynamics.sensors.seeker.seeker.bias
-=====================================
-
-.. currentmodule:: c4dynamics.sensors.seeker
-
-.. autoproperty:: seeker.bias
