+++ docs/html/_sources/api/generated/ekf/c4dynamics.filters.ekf.ekf.update.rst
@@ -1,6 +1,0 @@
-c4dynamics.filters.ekf.ekf.update
-=================================
-
-.. currentmodule:: c4dynamics.filters.ekf
-
-.. automethod:: ekf.update
