+++ docs/html/_sources/api/generated/yolov3/c4dynamics.detectors.yolo3_opencv.yolov3.detect.rst
@@ -1,6 +1,0 @@
-c4dynamics.detectors.yolo3\_opencv.yolov3.detect
-================================================
-
-.. currentmodule:: c4dynamics.detectors.yolo3_opencv
-
-.. automethod:: yolov3.detect
