+++ docs/html/_sources/api/generated/rb/c4dynamics.states.lib.rigidbody.rigidbody.ang_rates.rst
@@ -1,6 +1,0 @@
-c4dynamics.states.lib.rigidbody.rigidbody.ang\_rates
-====================================================
-
-.. currentmodule:: c4dynamics.states.lib.rigidbody
-
-.. autoproperty:: rigidbody.ang_rates
