+++ docs/html/_sources/api/generated/radar/c4dynamics.sensors.radar.radar.bias.rst
@@ -1,6 +1,0 @@
-c4dynamics.sensors.radar.radar.bias
-===================================
-
-.. currentmodule:: c4dynamics.sensors.radar
-
-.. autoproperty:: radar.bias
