+++ docs/html/_sources/api/generated/state/c4dynamics.states.state.state.plot.rst
@@ -1,6 +1,0 @@
-c4dynamics.states.state.state.plot
-==================================
-
-.. currentmodule:: c4dynamics.states.state
-
-.. automethod:: state.plot
