+++ docs/html/_sources/api/generated/seeker/c4dynamics.sensors.seeker.seeker.scale_factor.rst
@@ -1,6 +1,0 @@
-c4dynamics.sensors.seeker.seeker.scale\_factor
-==============================================
-
-.. currentmodule:: c4dynamics.sensors.seeker
-
-.. autoproperty:: seeker.scale_factor
