### 2024-12-23 f0e7d42
fix error in ekf docs, add paper.md, update dof6sim example
