### 2024-12-23 c9ee188
output paper path
