### 2025-01-09 b910cc4
user guide: extended kalman filter example
