### 2025-01-09 2f5f307
user guide: extended kalman filter example
