+++ docs/_sources/programs/index copy.rst
@@ -1,38 +1,0 @@
-Programs
-========
-
-
-.. toctree::
-   :caption: Getting Started
-   :hidden:
-   :maxdepth: 1
-
-   drifting_car
-   detect_track.ipynb
-   dof6sim.ipynb
-   ballisticestimation
-
-
-
-Selected examples that demonstrate 
-the framework's versatility, showcasing its use in diverse 
-fields such as object detection and tracking systems, 
-six degrees of freedom simulation, and the estimation of 
-variables for radar-tracked targets. 
-
-
-.. list-table:: 
-  :header-rows: 0
-
-  * - `Object Tracking <drifting_car>`_
-    - Object detection and tracking using YOLOv3 and Kalman filter.
-  * - `Multiple Tracking <detect_track.ipynb>`_
-    - Simultaneous objects detection object and tracking.
-  * - `6DOF <dof6sim.ipynb>`_
-    - Six degrees of freedom simulation of a missile pursuing a target.
-
-
-  * - `Ballistics Estimation <ballistic_coefficient.ipynb>`_
-    - Ballistic coefficient estimation using Extended Kalman fitler
-
-
