### 2025-01-30 f1a20cb
fix bug in the update equation in the kalman intro page
