+++ docs/source/api/index.rst
@@ -53,6 +53,7 @@ and saved to the cache for later usa.
 For more details, see the :mod:`datasets <c4dynamics.datasets>` module.
 
 
+
 Modules
 -------
 
