### 2025-01-30 eb02c4a
fix bug in the update equation in the kalman intro page
