### 2025-02-05 #224 @coatless
Hi Ziv/@C4dynamics thank you for making this presubmission inquiry. Sorry we haven't gotten back to you sooner as we were on break and undergoing an EiC transition.

In terms of scope, we'll need to consider the package underneath `Data processing and munging` category:

https://www.pyopensci.org/software-peer-review/about/package-scope.html#data-processing-and-munging

I'm working on broadening our categories; but, that will be the primary one used for this submission.

We'll also consider this package to be "in-scope" under a specialized field instead of a general "core" package (due to the new data structures).

Regarding other Python packages, I suppose we have:

- For event-driven or discrete event simulation, [`statepy`](https://pypi.org/project/StatePy/) [defunct], [`statemachine`](https://python-statemachine.readthedocs.io/en/latest/readme.html), or [`transition`](https://github.com/pytransitions/transitions).
- For decision-making or reinforcement learning systems, [`POMDPy`](https://github.com/pemami4911/POMDPy) (not maintained) and tools like `transition` (for FSM).

In comparison, this package presents a novel state interface and individual modules tailored for sensors, detectors, and filters.

----

Before submitting for review, please make sure to enable CI testing:

https://www.pyopensci.org/python-package-guide/tests/tests-ci.html

Please add to the README: Badges for the packages current published version, documentation and test suite build. (OPTIONAL: test coverage)

https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html

Please make sure to use capitalized names for `README.md`, `CODE_OF_CONDUCT.md`, and `CONTRIBUTING.md`.

Please add a `CHANGELOG.md` file discussing changes:

https://www.pyopensci.org/python-package-guide/documentation/repository-files/changelog-file.html
