### 2025-02-20 5d83d4f
add run-tests for CI
