### 2025-02-20 0b5bc06
add run-tests for CI
