### 2025-02-21 c898ad2
limit tests to python 3.11
