### 2025-02-21 8ee625a
limit tests to python 3.11
