### 2025-02-21 5520a2f
rename run tests yaml & add changelog
