+++ README.md
@@ -53,17 +53,17 @@ while maintaining flexibility and scalability.
 ## Installation 
 
 
-* PIP  
+* [PIP](https://pypi.org/project/c4dynamics/)
 
 ```
 >>> pip install c4dynamics
 ```
-[PyPi](https://pypi.org/project/c4dynamics/)
 
 
-* GitHub  
 
-To run the latest GitHub version, download [c4dynamics](https://github.com/C4dynamics/C4dynamics)
+* [GitHub](https://github.com/C4dynamics/C4dynamics)  
+
+To run the latest GitHub version, download 
 
 Install required packages:
 
