### 2025-02-23 #224 @c4dynamics
Hi,
I'm sorry.
I now see that the tests in the CI workflow are not really performed ("run 0 tests in 0sec").

Let me a few days to see what's going on and I'll update you.
