### 2025-02-23 4cdd2e8
mock open3d to avoid installation
