+++ .github/workflows/run-tests.yml
@@ -33,6 +33,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt  
+          pip install open3d  
 
       - name: List installed packages
         run: pip list
