### 2025-02-24 e36ebe0
skif tests if not display ui
