+++ paper/paper.md
@@ -10,6 +10,7 @@ tags:
 authors:
   - name: Ziv Meri
     orcid: 0000-0002-7932-0562
+    affiliation: C4DYNAMICS
 date: December 17 2024
 ---
 
