### 2025-03-06 #224 @coatless
@C4dynamics thank you for addressing those points. I'd like to move this from a `pre-submission` inquiry into a formal review inquiry.

This means that I'll need you to create a new issue on the repository that follows our formal software request instead of a pre-submission note.

https://github.com/pyOpenSci/software-submission/issues/new?template=submit-software-for-review.md

You can see the next steps in the process under the author's guide:

https://www.pyopensci.org/software-peer-review/how-to/author-guide.html#submit-your-package-for-peer-review

If you any more pre-submission inquiries, please let me know!
