+++ docs/source/programs/dof6_modules.py
@@ -1,116 +1,0 @@
-import numpy as np
-
-
-class control_system:
-    
-  Gn = 250 # gain factor relating aoa of ctrl surface to acc cmd per unit dynamic pressure, [rad*Pa/(m/s)^2]
-  afp = 0
-  afy = 0
-  tau = 0.04 
-  dt = 0
-
-  def __init__(self, dt, **kwargs):
-    self.dt = dt
-    self.__dict__.update(kwargs)
-
-  def update(self, ab_cmd, Q):
-    afp = -self.Gn * ab_cmd[2] / Q
-    afy =  self.Gn * ab_cmd[1] / Q
-    
-    self.afp = self.afp * np.exp(-self.dt / self.tau) + afp * (1 - np.exp(-self.dt / self.tau))
-    self.afy = self.afy * np.exp(-self.dt / self.tau) + afy * (1 - np.exp(-self.dt / self.tau))
-    
-    if abs(self.afp) > np.deg2rad(20):
-      self.afp = np.sign(self.afp) * np.deg2rad(20)
-    if abs(self.afy) > np.deg2rad(20):
-      self.afy = np.sign(self.afy) * np.deg2rad(20)
-    
-    return self.afp, self.afy 
-
-
-class aerodynamics():
-
-  s    = 0.0127
-  d    = 0.127
-  xref = 1.35
-  
-  mach_table = (0, 0.8, 1.14, 1.75, 2.5, 3.5) 
-  cD0_table  = (0.8, 0.8, 1.2, 1.15, 1.05, 0.94)
-  cLa_table  = (38,  39,  56,  55,   40,   33)
-  cMa_table  = (-160, -170, -185, -235, -190, -150) 
-  cMd_table  = (180, 250, 230, 130, 80, 45)
-  cMqcMadot_table = (-6000, -13000, -16000, -13500, -10000, -6000) 
-  k_table    = (0.0255, 0.0305, 0.0361, 0.0441, 0.0540, 0.0665)
-  
-  alt_table = (0, 2000, 4000, 6000)
-  pressure_table = (101325, 79501, 61660, 47217)
-  density_table  = (1.225, 1.0066, 0.81935, 0.66011)
-  speed_of_sound_table = (340.29, 332.53, 324.59, 316.45)
-
-  def f_coef(self, mach, alpha_total):
-    cLa = np.interp(mach, self.mach_table, self.cLa_table)
-    cD0 = np.interp(mach, self.mach_table, self.cD0_table)
-    k = np.interp(mach, self.mach_table, self.k_table)
-            
-    # lift and drag
-    cL = cLa * alpha_total
-    cD = cD0 + k * cL**2
-    
-    return cL, cD
-
-  def m_coef(self, mach, alpha, beta
-                , d_pitch, d_yaw, xcm 
-                  , Q, v, fAby, fAbz, q, r):
-      
-      cNb = cMa = np.interp(mach, self.mach_table, self.cMa_table)
-      cNd = cMd = np.interp(mach, self.mach_table, self.cMd_table)
-      cNrcNbdot = cMqcMadot = np.interp(mach, self.mach_table, self.cMqcMadot_table)
-      
-      # pitch and yaw moments 
-      # yb, zb normal force aero coefficient 
-      cNy = fAby / Q / self.s
-      cNz = fAbz / Q / self.s
-      cMref = cMa * alpha + cMd * d_pitch
-      cNref = cNb * beta  + cNd * d_yaw
-
-      # to center of mass
-      cM = cMref - cNz * (xcm - self.xref) / self.d + self.d / (2 * v) * cMqcMadot * q
-      cN = cNref - cNy * (xcm - self.xref) / self.d + self.d / (2 * v) * cNrcNbdot * r
-
-      return cM, cN
-
[…]
