+++ docs/source/conf.py
@@ -1,391 +1,0 @@
-# from IPython.display import display_html
-# display_html("<script>Jupyter.notebook.kernel.restart()</script>", raw = True)
-
-import re
-import os, sys
-# import importlib
-
-'''
-
-required packages:
-
-pip install sphinx
-pip install sphinx_design
-pip install sphinx_book_theme
-
-https://github.com/numpy/numpy/issues/22408
-Too light color for the label of "dev" version selection button
-The label of "dev" version selection button is colored light gray in the dark mode and the button itself is colored orange. The difference in brightness between those two colors is too small to read.
-In the light mode, the label color is black and it looks clear.
-
-ss 2022-10-08 15 55 00
-
-This is because pydata-sphinx-theme defines color for #version_switcher_button, whose value is #c9d1d9 in the dark mode although it is #333333 in the light mode. numpy.css's definition only overwrites the background color for "dev".
-
-H3-level section heading colors
-I think H3-level section heading colors is a bit too dark, but this may be a matter of opinion:
-
-ss 2022-10-08 15 54 25
-
-The cause is the same as that of the issue with the body text color; the color of h3 is overwritten in numpy.css.
-
-
-
-
-Each introduction page on your GitHub page, website, 
-and documentation serves different purposes and targets 
-different audiences. Here's a breakdown of the intention and focus for each:
-
-https://chatgpt.com/share/671776d9-c600-8002-bf9e-c9345a1fa792
-
-
-### 1. **GitHub Page (Repository README)**  
-**Purpose**: To introduce your project to developers and contributors.  
-**Focus**: Provide a concise overview of the project, its purpose, and how to get started.  
-**What to include**:  
-   - **Project Name** and **Brief Description**: Explain the problem your project solves or the features it offers.
-   - **Key Features**: Highlight what makes your project unique or useful.
-   - **Installation Instructions**: How someone can clone the repository and run the project locally.
-   - **Usage**: Basic examples of how to use the project or API.
-   - **Contributing Guidelines**: Link to guidelines for developers who want to contribute.
-   - **Links to Documentation**: Provide a link to the full documentation page.
-   - **License**: The open-source license under which your project is distributed.
-
-   readme: 
-    
-  ## Features 
-
-
-
-  ## Architecture
-
-  ## Roadmap 
-
-
-
-  ### 1. **GitHub Page (Repository README)**  
-  **Purpose**: To introduce your project to developers and contributors.  
-
-  **Focus**: Provide a concise overview of the project, its purpose, and how to get started.  
-
-  **What to include**:  
-
-   - **Project Name** and **Brief Description**: 
-   Explain the problem your project solves or the features it offers.
-
-   - **Key Features**: 
-   Highlight what makes your project unique or useful.
-
-   - **Installation Instructions**: 
-   How someone can clone the repository and run the project locally.
-   - **Usage**: Basic examples of how to use the project or API.
-   
-   - **Contributing Guidelines**: Link to guidelines for developers who want to contribute.
-   - **Links to Documentation**: Provide a link to the full documentation page.
-   - **License**: The open-source license under which your project is distributed.
-
-
-### 2. **Website (General Audience)**
-**Purpose**: To introduce your product or project to a broad, often non-technical audience.
-**Focus**: Showcase the benefits and value of your project in a way that appeals to potential users, stakeholders, or customers.
[…]
