+++ docs/source/api/generated/kalman/c4dynamics.filters.kalman.kalman.update.rst
@@ -1,6 +1,0 @@
-c4dynamics.filters.kalman.kalman.update
-=======================================
-
-.. currentmodule:: c4dynamics.filters.kalman
-
-.. automethod:: kalman.update
