+++ docs/source/api/detectors.yolov3.rst
@@ -1,24 +1,0 @@
-YOLOv3 
-======
-.. currentmodule:: c4dynamics.detectors.yolo3_opencv
-
-
-.. autoclass:: yolov3 
-
-
-   **Properties**
-
-   .. autosummary:: 
-      :toctree: generated/yolov3
-
-      yolov3.nms_th
-      yolov3.confidence_th
-
-
-   **Methods**
-
-   .. autosummary:: 
-      :toctree: generated/yolov3
-
-      yolov3.detect
-
