+++ docs/source/api/generated/state/c4dynamics.states.state.state.X.rst
@@ -1,6 +1,0 @@
-c4dynamics.states.state.state.X
-===============================
-
-.. currentmodule:: c4dynamics.states.state
-
-.. autoproperty:: state.X
