+++ docs/source/api/generated/datasets/c4dynamics.datasets.d3_model.rst
@@ -1,6 +1,0 @@
-c4dynamics.datasets.d3\_model
-=============================
-
-.. currentmodule:: c4dynamics.datasets
-
-.. autofunction:: d3_model
