+++ docs/source/api/generated/state/c4dynamics.states.state.state.X0.rst
@@ -1,6 +1,0 @@
-c4dynamics.states.state.state.X0
-================================
-
-.. currentmodule:: c4dynamics.states.state
-
-.. autoproperty:: state.X0
