+++ docs/source/api/generated/eqm/c4dynamics.eqm.integrate.int6.rst
@@ -1,6 +1,0 @@
-c4dynamics.eqm.integrate.int6
-=============================
-
-.. currentmodule:: c4dynamics.eqm.integrate
-
-.. autofunction:: int6
