+++ docs/source/api/generated/eqm/c4dynamics.eqm.derivs.eqm3.rst
@@ -1,6 +1,0 @@
-c4dynamics.eqm.derivs.eqm3
-==========================
-
-.. currentmodule:: c4dynamics.eqm.derivs
-
-.. autofunction:: eqm3
