+++ docs/source/api/generated/ekf/c4dynamics.filters.ekf.ekf.predict.rst
@@ -1,6 +1,0 @@
-c4dynamics.filters.ekf.ekf.predict
-==================================
-
-.. currentmodule:: c4dynamics.filters.ekf
-
-.. automethod:: ekf.predict

[…]
