+++ docs/_sources/installation.rst
@@ -1,0 +1,46 @@
+Installation
+============
+
+
+pip  
+---
+
+The package is available on PyPI. Install the latest stable version using:
+
+.. code-block::
+
+  pip install c4dynamics 
+
+For more details, visit: https://pypi.org/project/c4dynamics/
+
+
+GitHub
+------
+
+Download the latest version (not stable) for development: 
+
+.. code-block::
+
+  git clone https://github.com/C4dynamics/C4dynamics
+
+
+Install the required packages:
+
+.. code-block::
+
+  pip install -r requirements.txt
+
+
+ 
+Import 
+------
+
+Most examples import c4dynamics and access its 
+modules and utils using the alias c4d:
+
+
+.. code-block:: 
+  
+  import c4dynamics as c4d 
+
+
