+++ docs/_sources/api/generated/datasets/c4dynamics.datasets.download_all.rst
@@ -1,0 +1,6 @@
+c4dynamics.datasets.download\_all
+=================================
+
+.. currentmodule:: c4dynamics.datasets
+
+.. autofunction:: download_all
