+++ docs/_sources/api/generated/rb/c4dynamics.states.lib.rigidbody.rigidbody.animate.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.lib.rigidbody.rigidbody.animate
+=================================================
+
+.. currentmodule:: c4dynamics.states.lib.rigidbody
+
+.. automethod:: rigidbody.animate
