+++ docs/_sources/programs/index.rst
@@ -1,0 +1,32 @@
+User Guide
+==========
+
+
+.. toctree::
+  :caption: Getting Started
+  :hidden:
+  :maxdepth: 1
+
+  dof6sim.ipynb
+  ballistic_coefficient.ipynb
+  car_tracker.ipynb
+
+  
+
+
+
+Selected programs that demonstrate the framework's versatility, 
+showcasing its use in diverse fields. 
+
+
+.. list-table:: 
+  :header-rows: 0
+
+  * - `6DOF <dof6sim.ipynb>`_
+    - Six degrees of freedom simulation of a missile pursuing a target
+  * - `EKF <ballistic_coefficient.ipynb>`_
+    - Ballistic coefficient estimation using Extended Kalman fitler
+  * - `Tracker <car_tracker.ipynb>`_
+    - Use a Kalman filter to track a vehicle across video frames.
+
+
