+++ docs/_sources/api/sensors.radar.rst
@@ -1,0 +1,25 @@
+Radar 
+=====
+.. currentmodule:: c4dynamics.sensors.radar
+
+
+.. autoclass:: radar 
+
+
+   **Properties**
+
+   .. autosummary:: 
+      :toctree: generated/radar
+
+      radar.bias
+      radar.scale_factor
+
+
+   **Methods**
+
+   .. autosummary:: 
+      :toctree: generated/radar
+
+      radar.measure
+
+
