+++ docs/_sources/api/generated/rb/c4dynamics.states.lib.rigidbody.rigidbody.BR.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.lib.rigidbody.rigidbody.BR
+============================================
+
+.. currentmodule:: c4dynamics.states.lib.rigidbody
+
+.. autoproperty:: rigidbody.BR
