+++ docs/_sources/api/generated/state/c4dynamics.states.state.state.X.rst
@@ -1,0 +1,6 @@
+c4dynamics.states.state.state.X
+===============================
+
+.. currentmodule:: c4dynamics.states.state
+
+.. autoproperty:: state.X
