+++ docs/_sources/api/generated/utils/c4dynamics.utils.cprint.cprint.rst
@@ -1,0 +1,6 @@
+c4dynamics.utils.cprint.cprint
+==============================
+
+.. currentmodule:: c4dynamics.utils.cprint
+
+.. autofunction:: cprint
