+++ docs/_sources/api/generated/radar/c4dynamics.sensors.radar.radar.scale_factor.rst
@@ -1,0 +1,6 @@
+c4dynamics.sensors.radar.radar.scale\_factor
+============================================
+
+.. currentmodule:: c4dynamics.sensors.radar
+
+.. autoproperty:: radar.scale_factor
