+++ docs/_sources/api/generated/rotmat/c4dynamics.rotmat.animate.animate.rst
@@ -1,0 +1,6 @@
+c4dynamics.rotmat.animate.animate
+=================================
+
+.. currentmodule:: c4dynamics.rotmat.animate
+
+.. autofunction:: animate
